requirements.txt
matplotlib==3.10.8
numpy==2.4.3
Pillow==12.1.1
thop==0.1.1.post2209072238
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
